requirements.txt
langchain-community
langchain-chroma
langchain-huggingface
PyMuPDF
PyMuPDF4LLM
PyYAML
sentence-transformers
rich
readchar